+++ scripts/stubs/react-devtools-core-pkg/index.js
@@ -1,0 +1,6 @@
+const noop = () => {};
+
+export default {
+  initialize: noop,
+  connectToDevTools: noop,
+};